// frontend/src/app/models/i-personas.ts
export interface IPersonas {
  _id?: string;
  email?: string;
  apellido?: string;
  nombre?: string;
  name?: string;
  cuit?: string;
  fijo?: string;
  celular?: string;
  direccion?: string;
  localidad?: string;
  provincia?: string;
  zipcode?: string;
  pais?: string;
  coeficiente?: number;
  categoria?: [];
}